requirements.txt
# FastAPI Agent Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Authentication
pyjwt[crypto]==2.8.0
cryptography>=43.0.0
python-jose[cryptography]==3.3.0

# Database
supabase==2.0.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23

# AI Orchestration Frameworks (Core)
openai==1.84.0
anthropic==0.25.0

# AI Frameworks (Optional - will gracefully fallback if not available)
praisonaiagents>=0.0.70
crewai>=0.70.0
pyautogen>=0.2.30

# HTTP & API utilities
httpx==0.24.1
aiohttp>=3.10.0
requests==2.31.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-multipart==0.0.6 